audit-automator/requirements.txt
# GCP and Vertex AI
google-cloud-storage
google-cloud-aiplatform
google-cloud-documentai

# For local development, to load .env files
python-dotenv
jsonschema # For validating AI model outputs
PyMuPDF # For PDF processing (provides 'fitz' module)